# src/main/resources/static/openapi.yaml
...
openapi: 3.0.1
info:
  title: "Channel-Talk Notion Integration API"
  description: "채널톡과 노션을 연동하는 서비스 API 명세"
  version: "1.0.0"

# ===============================================
# 1. API Paths (엔드포인트 정의)
# ===============================================
paths:
...
  /integrations/notion/callback:
    get:
      summary: "노션 Oauth 콜백"
      description: "노션 Oauth 승인 후, 리디렉션되는 콜백 URL입니다. 서버는 토큰을 발급받아 저장합니다."
      tags: ["IntegrATIONS"]
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: "노션에서 발급한 1회용 인증 코드"
        - in: query
          name: error
          schema:
            type: string
          required: false
          description: "인증 실패 시 (예: access_denied)"
      responses:
        '200':
          description: "연동 성공. (성공 페이지로 리디렉션 또는 스크립트)"
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>연동에 성공했습니다. 창을 닫아주세요.</body></html>"
        '400':
          description: "유효하지 않은 요청 (code 누락 등)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
                example:
                  error: "INVALID_NOTION_CODE"
                  message: "유효하지 않은 노션 인증 코드입니다."
  /integrations/notion/status:
    get:
      summary: "노션 연동 상태 조회"
      description: "현재 노션 연동 상태, 연결된 워크스페이스 이름, 링크 등을 조회합니다."
      tags: ["Integrations"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "연동 상태 조회 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionStatusResponse"
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "노션이 아직 연동되지 않았습니다."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
                example:
                  error: "NOTION_NOT_CONNECTED"
                  message: "노션이 아직 연동되지 않았습니다."
...
  /notion/databases:
    post:
      summary: "노션 데이터베이스 생성"
      description: |
        노션의 특정 페이지 하위에 새 데이터베이스를 생성합니다.
        (주로 Oauth 연동 직후, 최초 1회 설정 시 사용됩니다.)
      tags: ["Notion"]
      security:
        - bearerAuth: []
      requestBody:
        description: "생성할 데이터베이스의 정보"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotionDatabaseRequest"
      responses:
        '201':
          description: "데이터베이스 생성 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionDatabaseResponse"
        '400':
          description: "잘못된 요청 (예: parentPageId가 유효하지 않음)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "상위 페이지(parentPage)를 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notion/pages/{pageId}:
    get:
      summary: "노션 페이지 조회"
      description: "특정 노션 페이지의 정보를 조회합니다."
      tags: ["Notion"]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pageId
          schema:
            type: string
          required: true
          description: "조회할 노션 페이지 ID"
      responses:
        '200':
          description: "페이지 조회 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionPageResponse"
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "페이지를 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: "노션 페이지 수정"
      description: "특정 노션 페이지의 속성을 수정합니다."
      tags: ["Notion"]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pageId
          schema:
            type: string
          required: true
          description: "수정할 노션 페이지 ID"
      requestBody:
        description: "수정할 페이지 속성"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotionPageRequest"
            example:
              properties:
                title:
                  title:
                    - text:
                        content: "업데이트된 제목"
      responses:
        '200':
          description: "페이지 수정 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionPageResponse"
        '400':
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "페이지를 찾을 수 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notion/pages:
    post:
      summary: "노션 페이지 생성"
      description: "새로운 노션 페이지를 생성합니다."
      tags: ["Notion"]
      security:
        - bearerAuth: []
      requestBody:
        description: "생성할 페이지 정보"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotionPageRequest"
            example:
              parent:
                type: "page_id"
                page_id: "id"
              properties:
                title:
                  title:
                    - text:
                        content: "생성한 페이지"
              children:
                - object: "block"
                  type: "paragraph"
                  paragraph:
                    rich_text:
                      - type: "text"
                        text:
                          content: "제목"
      responses:
        '201':
          description: "페이지 생성 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotionPageResponse"
        '400':
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------
  # /webhooks
  # ---------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    channelTalkSignature:
      type: apiKey
      in: header
      name: X-Channel-Signature
      description: "채널톡 웹훅의 유효성 검증을 위한 시그니처"

  # ---------------------------------------------
  # Schemas (Data Models)
  # ---------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      description: "공통 에러 응답 모델"
      properties:
        error:
          type: string
          description: "에러 코드"
        message:
          type: string
          description: "에러 상세 메시지"

    # Auth
...
    NotionStatusResponse:
      type: object
      description: "노션 연동 상태 응답"
      properties:
        connected:
          type: boolean
          example: true
        workspaceName:
          type: string
          example: "홍길동의 워크스페이스"
        workspaceUrl:
          type: string
          format: uri
          example: "https://www.notion.so/..."
      required:
        - connected

    # Statistics
...
    NotionDatabaseRequest:
      type: object
      description: "노션 DB 생성 요청"
      properties:
        parentPageId:
          type: string
          description: "데이터베이스가 생성될 상위 페이지의 ID"
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          description: "새 데이터베이스의 제목"
          example: "고객 문의 아카이브"
      required:
        - parentPageId
        - title
    NotionDatabaseResponse:
      type: object
      description: "노션 DB 생성 응답"
      properties:
        databaseId:
          type: string
          description: "생성된 데이터베이스의 ID"
          example: "789e4567-e89b-12d3-a456-426614174aaa"
        databaseUrl:
          type: string
          format: uri
          description: "생성된 데이터베이스의 URL"
          example: "https://www.notion.so/..."
      required:
        - databaseId
        - databaseUrl
    NotionPageRequest:
      type: object
      description: "노션 페이지(행) 생성 요청"
      properties:
        properties:
          type: object
          description: "노션 DB의 속성(컬럼)에 맞게 매핑된 데이터"
          example: {
            "문의일시": { "date": { "start": "2025-11-06T13:00:00Z" } },
            "고객명": { "title": [ { "text": { "content": "홍길동" } } ] },
            "태그": { "multi_select": [ { "name": "환불" }, { "name": "긴급" } ] }
          }
        content:
          type: array
          description: "페이지 본문에 추가될 블록(채팅 내용)"
          items:
            type: object
          example: [
            {
              "object": "block",
              "type": "heading_2",
              "heading_2": { "rich_text": [ { "text": { "content": "채팅 내용" } } ] }
            },
            {
              "object": "block",
              "type": "paragraph",
              "paragraph": { "rich_text": [ { "text": { "content": "고객: 안녕하세요..." } } ] }
            }
          ]
      required:
        - properties
    NotionPageResponse:
      type: object
      description: "노션 페이지(행) 생성 응답"
      properties:
        pageId:
          type: string
          description: "생성된 페이지의 ID"
          example: "abc-123..."
        pageUrl:
          type: string
          format: uri
          description: "생성된 페이지의 URL"
          example: "https://www.notion.so/..."
      required:
        - pageId
        - pageUrl